backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
torch==2.1.0
numpy==1.24.3
pandas==2.1.0
scikit-learn==1.3.0
huggingface-hub==0.25.2
python-dotenv==1.0.0
gunicorn==21.2.0
openai==1.3.0
